sam-python-learning/src/auth/requirements.txt
PyJWT>=2.8.0
boto3